fix the hot region API cannot work when hot scheduler is disabled
