+++ server/schedulers/utils.go
@@ -15,7 +15,6 @@
 package schedulers
 
 import (
-	"fmt"
 	"net/url"
 	"strconv"
 	"time"
@@ -291,250 +290,6 @@ func diffCmp(ldCmp storeLoadCmp) storeLPCmp {
 	return func(lp1, lp2 *statistics.StoreLoadPred) int {
 		return ldCmp(lp1.Diff(), lp2.Diff())
 	}
-}
-
-// storeCollector define the behavior of different engines of stores.
-type storeCollector interface {
-	// Engine returns the type of Store.
-	Engine() string
-	// Filter determines whether the Store needs to be handled by itself.
-	Filter(info *statistics.StoreSummaryInfo, kind core.ResourceKind) bool
-	// GetLoads obtains available loads from storeLoads and peerLoadSum according to rwTy and kind.
-	GetLoads(storeLoads, peerLoadSum []float64, rwTy statistics.RWType, kind core.ResourceKind) (loads []float64)
-}
-
-type tikvCollector struct{}
-
-func newTikvCollector() storeCollector {
-	return tikvCollector{}
-}
-
-func (c tikvCollector) Engine() string {
-	return core.EngineTiKV
-}
-
-func (c tikvCollector) Filter(info *statistics.StoreSummaryInfo, kind core.ResourceKind) bool {
-	if info.IsTiFlash {
-		return false
-	}
-	switch kind {
-	case core.LeaderKind:
-		return info.Store.AllowLeaderTransfer()
-	case core.RegionKind:
-		return true
-	}
-	return false
-}
-
-func (c tikvCollector) GetLoads(storeLoads, peerLoadSum []float64, rwTy statistics.RWType, kind core.ResourceKind) (loads []float64) {
-	loads = make([]float64, statistics.DimLen)
-	switch rwTy {
-	case statistics.Read:
-		loads[statistics.ByteDim] = storeLoads[statistics.StoreReadBytes]
-		loads[statistics.KeyDim] = storeLoads[statistics.StoreReadKeys]
-		loads[statistics.QueryDim] = storeLoads[statistics.StoreReadQuery]
-	case statistics.Write:
-		switch kind {
-		case core.LeaderKind:
-			// Use sum of hot peers to estimate leader-only byte rate.
-			// For write requests, Write{Bytes, Keys} is applied to all Peers at the same time,
-			// while the Leader and Follower are under different loads (usually the Leader consumes more CPU).
-			// Write{QPS} does not require such processing.
-			loads[statistics.ByteDim] = peerLoadSum[statistics.ByteDim]
-			loads[statistics.KeyDim] = peerLoadSum[statistics.KeyDim]
-			loads[statistics.QueryDim] = storeLoads[statistics.StoreWriteQuery]
-		case core.RegionKind:
-			loads[statistics.ByteDim] = storeLoads[statistics.StoreWriteBytes]
-			loads[statistics.KeyDim] = storeLoads[statistics.StoreWriteKeys]
-			// The `write-peer` does not have `QueryDim`
-		}
-	}
-	return
-}
-
-type tiflashCollector struct {
-	isTraceRegionFlow bool
-}
-
-func newTiFlashCollector(isTraceRegionFlow bool) storeCollector {
-	return tiflashCollector{isTraceRegionFlow: isTraceRegionFlow}
-}
-
-func (c tiflashCollector) Engine() string {
-	return core.EngineTiFlash
-}
-
-func (c tiflashCollector) Filter(info *statistics.StoreSummaryInfo, kind core.ResourceKind) bool {
-	switch kind {
-	case core.LeaderKind:
-		return false
-	case core.RegionKind:
-		return info.IsTiFlash
-	}
-	return false
-}
-
-func (c tiflashCollector) GetLoads(storeLoads, peerLoadSum []float64, rwTy statistics.RWType, kind core.ResourceKind) (loads []float64) {
-	loads = make([]float64, statistics.DimLen)
-	switch rwTy {
-	case statistics.Read:
-		// TODO: Need TiFlash StoreHeartbeat support
-	case statistics.Write:
-		switch kind {
-		case core.LeaderKind:
-			// There is no Leader on TiFlash
-		case core.RegionKind:
-			// TiFlash is currently unable to report statistics in the same unit as Region,
-			// so it uses the sum of Regions. If it is not accurate enough, use sum of hot peer.
-			if c.isTraceRegionFlow {
-				loads[statistics.ByteDim] = storeLoads[statistics.StoreRegionsWriteBytes]
-				loads[statistics.KeyDim] = storeLoads[statistics.StoreRegionsWriteKeys]
-			} else {
-				loads[statistics.ByteDim] = peerLoadSum[statistics.ByteDim]
-				loads[statistics.KeyDim] = peerLoadSum[statistics.KeyDim]
-			}
-			// The `wite-peer` does not have `QueryDim`
-		}
-	}
-	return
-}
-
-// summaryStoresLoad Load information of all available stores.
-// it will filter the hot peer and calculate the current and future stat(rate,count) for each store
-func summaryStoresLoad(
-	storeInfos map[uint64]*statistics.StoreSummaryInfo,
-	storesLoads map[uint64][]float64,
-	storeHotPeers map[uint64][]*statistics.HotPeerStat,
-	isTraceRegionFlow bool,
-	rwTy statistics.RWType,
-	kind core.ResourceKind,
-) map[uint64]*statistics.StoreLoadDetail {
-	// loadDetail stores the storeID -> hotPeers stat and its current and future stat(rate,count)
-	loadDetail := make(map[uint64]*statistics.StoreLoadDetail, len(storesLoads))
-
-	tikvLoadDetail := summaryStoresLoadByEngine(
-		storeInfos,
-		storesLoads,
-		storeHotPeers,
-		rwTy, kind,
-		newTikvCollector(),
-	)
-	tiflashLoadDetail := summaryStoresLoadByEngine(
-		storeInfos,
-		storesLoads,
-		storeHotPeers,
-		rwTy, kind,
-		newTiFlashCollector(isTraceRegionFlow),
-	)
-
-	for _, detail := range append(tikvLoadDetail, tiflashLoadDetail...) {
-		loadDetail[detail.GetID()] = detail
-	}
-	return loadDetail
-}
-
-func summaryStoresLoadByEngine(
-	storeInfos map[uint64]*statistics.StoreSummaryInfo,
-	storesLoads map[uint64][]float64,
-	storeHotPeers map[uint64][]*statistics.HotPeerStat,
-	rwTy statistics.RWType,
-	kind core.ResourceKind,
-	collector storeCollector,
-) []*statistics.StoreLoadDetail {
-	loadDetail := make([]*statistics.StoreLoadDetail, 0, len(storeInfos))
-	allStoreLoadSum := make([]float64, statistics.DimLen)
-	allStoreCount := 0
-	allHotPeersCount := 0
-
-	for _, info := range storeInfos {
-		store := info.Store
-		id := store.GetID()
-		storeLoads, ok := storesLoads[id]
-		if !ok || !collector.Filter(info, kind) {
-			continue
-		}
-
-		// Find all hot peers first
-		var hotPeers []*statistics.HotPeerStat
-		peerLoadSum := make([]float64, statistics.DimLen)
-		// TODO: To remove `filterHotPeers`, we need to:
-		// HotLeaders consider `Write{Bytes,Keys}`, so when we schedule `writeLeader`, all peers are leader.
-		for _, peer := range filterHotPeers(kind, storeHotPeers[id]) {
-			for i := range peerLoadSum {
-				peerLoadSum[i] += peer.GetLoad(statistics.GetRegionStatKind(rwTy, i))
-			}
-			hotPeers = append(hotPeers, peer.Clone())
-		}
-		{
-			// Metric for debug.
-			ty := "byte-rate-" + rwTy.String() + "-" + kind.String()
-			hotPeerSummary.WithLabelValues(ty, fmt.Sprintf("%v", id)).Set(peerLoadSum[statistics.ByteDim])
-			ty = "key-rate-" + rwTy.String() + "-" + kind.String()
-			hotPeerSummary.WithLabelValues(ty, fmt.Sprintf("%v", id)).Set(peerLoadSum[statistics.KeyDim])
-			ty = "query-rate-" + rwTy.String() + "-" + kind.String()
-			hotPeerSummary.WithLabelValues(ty, fmt.Sprintf("%v", id)).Set(peerLoadSum[statistics.QueryDim])
-		}
-
-		loads := collector.GetLoads(storeLoads, peerLoadSum, rwTy, kind)
-		for i := range allStoreLoadSum {
-			allStoreLoadSum[i] += loads[i]
-		}
-		allStoreCount += 1
-		allHotPeersCount += len(hotPeers)
-
-		// Build store load prediction from current load and pending influence.
-		stLoadPred := (&statistics.StoreLoad{
-			Loads: loads,
-			Count: float64(len(hotPeers)),
-		}).ToLoadPred(rwTy, info.PendingSum)
-
-		// Construct store load info.
-		loadDetail = append(loadDetail, &statistics.StoreLoadDetail{
-			Info:     info,
-			LoadPred: stLoadPred,
-			HotPeers: hotPeers,
-		})
-	}
-
-	if allStoreCount == 0 {
-		return loadDetail
-	}
-
-	expectCount := float64(allHotPeersCount) / float64(allStoreCount)
-	expectLoads := make([]float64, len(allStoreLoadSum))
-	for i := range expectLoads {
-		expectLoads[i] = allStoreLoadSum[i] / float64(allStoreCount)
-	}
-	{
-		// Metric for debug.
-		engine := collector.Engine()
-		ty := "exp-byte-rate-" + rwTy.String() + "-" + kind.String()
-		hotPeerSummary.WithLabelValues(ty, engine).Set(expectLoads[statistics.ByteDim])
-		ty = "exp-key-rate-" + rwTy.String() + "-" + kind.String()
-		hotPeerSummary.WithLabelValues(ty, engine).Set(expectLoads[statistics.KeyDim])
-		ty = "exp-query-rate-" + rwTy.String() + "-" + kind.String()
-		hotPeerSummary.WithLabelValues(ty, engine).Set(expectLoads[statistics.QueryDim])
-		ty = "exp-count-rate-" + rwTy.String() + "-" + kind.String()
-		hotPeerSummary.WithLabelValues(ty, engine).Set(expectCount)
-	}
-	expect := statistics.StoreLoad{
-		Loads: expectLoads,
-		Count: float64(allHotPeersCount) / float64(allStoreCount),
-	}
-	for _, detail := range loadDetail {
-		detail.LoadPred.Expect = expect
-	}
-	return loadDetail
-}
-
-func filterHotPeers(kind core.ResourceKind, peers []*statistics.HotPeerStat) []*statistics.HotPeerStat {
-	ret := make([]*statistics.HotPeerStat, 0, len(peers))
-	for _, peer := range peers {
-		if kind != core.LeaderKind || peer.IsLeader() {
-			ret = append(ret, peer)
-		}
-	}
-	return ret
 }
 
 type retryQuota struct {
